requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.5
pydantic-core==2.41.5
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2
anthropic==0.75.0